backend/requirements.txt
# 基础工具
setuptools>=65.5.1
wheel>=0.38.4
datasets>=2.14.1
torch>=2.1.1
simplejson>=3.19.3
sortedcontainers>=2.4.0

# Web框架
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
starlette>=0.27.0
typing-extensions>=4.8.0

# 异步和并发
celery==5.3.4
redis==5.0.1
aiofiles==23.2.1
click>=8.1.7
billiard>=4.1.0
kombu>=5.3.2
amqp>=5.1.1

# 配置管理
pydantic-settings==2.0.3
pydantic>=2.4.2
python-dotenv==1.0.0

# 文件处理
PyPDF2==3.0.1
python-docx==0.8.11
ffmpeg-python==0.2.0

# 科学计算
numpy>=1.21.0

# 新增依赖
beautifulsoup4==4.12.2
requests==2.31.0
dashscope==1.22.1
